Rename .env to .env.example
It caught me by surprise during development that fusion was reading environment variables from .env, as that's not really clear anywhere.

I think it's a bit of a gotcha, so I added logging to make the behavior clearer, and I renamed the default .env file to .env.example so that the user has to explicitly choose to use a .env file.

Keeping the real .env file out of source control also protects developers from accidentally committing their real password to source control and leaking it by mistake.