Check for error on retrieving the session
We are currently ignoring the error when we retrieve the session for a given HTTP cookie. I don't see any reason to ignore the error, as the value of the Session object is undefined unless err == nil.

This change updates the code so that we return an error if session.Get returns a non-nil error code.