Add unit tests for pull.FetchFeed
I thought it would be helpful to get unit test coverage for the feed fetch logic.

Currently, our logic is pretty simple, but I'd like fusion to eventually be able to support the If-Modified-Since HTTP header, so we don't force servers to send the same data over and over. At that point, extra testing will be especially helpful, but I thought it would be a good idea to get testing infrastructure in place now.