Change FusionRequest to use a non-pointer type parameter (#106)
FusionRequest currently specifies model.FeedRequestOptions as a pointer rather than as a regular parameter. This is unnecessary, as it's easy for us to treat model.FeedRequestOptions{} as the 'default options' value. With it as a pointer, we clutter our code with extra != nil checks.

This change updates FusionRequest to just take a model.FeedRequestOptions rather than a *model.FeedRequestOptions.

Co-authored-by: rook1e <rook1e404@outlook.com>